requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.1.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
alembic==1.12.1
itsdangerous==2.1.2
MarkupSafe==2.1.3
